+++ lib/flutter_folderview.dart
@@ -21,3 +21,4 @@ export 'themes/folder_view_theme.dart';
 export 'themes/parent_node_theme.dart';
 export 'themes/node_tooltip_theme.dart';
 export 'widgets/folder_view.dart';
+export 'widgets/folder_view_content.dart' show isScaleModifierPressed;